requirements.txt
aiogram==3.3.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
solana==0.30.2
solders==0.18.1
base58==2.1.1
cryptography==41.0.7
aiofiles==23.2.1
pillow==10.1.0
pydantic==2.5.3
pydantic-settings==2.1.0

